+++ backend/src/controllers/authController.js
@@ -1,0 +1,19 @@
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcrypt');
+const usersService = require('../services/usersService');
+
+async function loginUser(req, res) {
+  const { email, password } = req.body;
+  const user = usersService.findUserByEmail(email);
+  if (!user) {
+    return res.status(401).json({ error: 'Invalid credentials' });
+  }
+  const validPassword = await bcrypt.compare(password, user.password);
+  if (!validPassword) {
+    return res.status(401).json({ error: 'Invalid credentials' });
+  }
+  const token = jwt.sign({ email }, process.env.JWT_SECRET, { expiresIn: '1h' });
+  res.json({ token });
+}
+
+module.exports = { loginUser };
