+++ backend/src/controllers/usersController.js
@@ -1,4 +1,3 @@
-const jwt = require('jsonwebtoken');
 const usersService = require('../services/usersService');
 
 async function createUser(req, res) {
@@ -16,15 +15,4 @@ function getUsers(req, res) {
   res.json(users);
 }
 
-function loginUser(req, res) {
-  const { email, password } = req.body;
-  const users = usersService.getUsers();
-  const user = users.find((u) => u.email === email && u.password === password);
-  if (!user) {
-    return res.status(401).json({ error: 'Invalid credentials' });
-  }
-  const token = jwt.sign({ email }, process.env.JWT_SECRET, { expiresIn: '1h' });
-  res.json({ token });
-}
-
-module.exports = { createUser, getUsers, loginUser };
+module.exports = { createUser, getUsers };
