+++ backend/src/routes/authRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { loginUser } = require('../controllers/usersController');
+const { loginUser } = require('../controllers/authController');
 
 router.post('/login', loginUser);
